requirements.txt
fastapi~=0.115.3
asyncssh~=2.18.0
motor~=3.6.0
uvloop~=0.21.0
aiohttp~=3.10.10
pydantic~=2.9.2
pydantic_core~=2.23.4
pymongo~=4.9.2
APScheduler~=3.10.4
asyncvnc~=1.3.0
pillow~=11.0.0